Create README.md
Last practice of the introduction to programming.
We will create a function that allows us to return an array with the mean of its neighbors. The process of transforming each point to the average of its neighbors is a very common image filter. The big difference is in the size of the images, represented as arrays of pixels (3 numbers representing the red, green, and blue components).